Move permission options to itw own class (#6490)
